pyproject.toml
[project]
name = "vercel-fastapi-starter"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.121.0",
  "uvicorn>=0.34.0",
  "supabase",
  "python-dotenv",
  "python-jose[cryptography]",
  "passlib[bcrypt]",
  "email-validator",
  "python-multipart",
  "cryptography",
]
